refactor: moves each use case to its own directory
