+++ src/use-cases/calculateMealTicket/CalculateMealTicket.ts
@@ -1,4 +1,4 @@
-import { FixDecimalCases } from '../utils/FixDecimalCases';
+import { FixDecimalCases } from '../../utils/FixDecimalCases';
 
 interface ICalculateMealTicketProps {
   employeePosition: string;
