+++ src/use-cases/calculateMinSalary/CalculateMinSalary.spec.ts
