+++ src/use-cases/getWorkCoefficient/GetWorkCoefficient.spec.ts
