+++ src/use-cases/getWorkCoefficient/GetWorkCoefficient.ts
